Add test for usd, taxes and total price
