+++ .gitignore
@@ -1,7 +1,6 @@
 .env
 
 
-
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
