+++ requirements.txt
@@ -3,3 +3,4 @@ sendgrid==5.6.0
 pandas
 gspread
 oauth2client
+pytest
